@@ -2,6 +2,7 @@
   "name": "kobber-wiki",
   "version": "1.0.0",
   "private": true,
+  "type": "module",
   "scripts": {
     "clean": "rm -rf .turbo node_modules .next",
     "dev": "next dev",
@@ -14,11 +15,13 @@
     "db:drop": "drizzle-kit drop",
     "db:studio": "drizzle-kit studio"
   },
-  "dependencies": {
+  "devDependencies": {
     "@auth/drizzle-adapter": "^1.7.2",
     "@gyldendal/kobber-base": "*",
     "@gyldendal/kobber-components": "*",
     "@gyldendal/kobber-icons": "*",
+    "@ianvs/prettier-plugin-sort-imports": "^4.3.1",
+    "@neondatabase/serverless": "^0.10.1",
     "@radix-ui/react-aspect-ratio": "^1.1.1",
     "@radix-ui/react-checkbox": "^1.1.3",
     "@radix-ui/react-collapsible": "^1.1.2",
@@ -29,37 +32,34 @@
     "@radix-ui/react-scroll-area": "^1.2.2",
     "@radix-ui/react-slot": "^1.1.1",
     "@radix-ui/react-tooltip": "^1.1.6",
-    "class-variance-authority": "^0.7.0",
-    "clsx": "^2.1.1",
-    "drizzle-orm": "^0.36.0",
-    "lucide-react": "^0.429.0",
-    "next": "^15.1.4",
-    "next-auth": "^5.0.0-beta.25",
-    "react": "^19.0.0",
-    "react-dom": "^19.0.0",
-    "tailwind-merge": "^2.6.0",
-    "tailwindcss-animate": "^1.0.7"
-  },
-  "devDependencies": {
-    "@ianvs/prettier-plugin-sort-imports": "^4.3.1",
-    "@neondatabase/serverless": "^0.10.1",
     "@tailwindcss/typography": "^0.5.14",
     "@types/node": "^20",
     "@types/react": "^19.0.6",
     "@types/react-dom": "^19.0.3",
-    "@typescript-eslint/parser": "^8.2.0",
+    "@typescript-eslint/parser": "^8.20.0",
     "autoprefixer": "^10.4.20",
+    "class-variance-authority": "^0.7.0",
+    "clsx": "^2.1.1",
     "drizzle-kit": "^0.27.1",
-    "eslint": "^8",
-    "eslint-config-next": "14.2.5",
-    "eslint-config-prettier": "^9.1.0",
+    "drizzle-orm": "^0.36.0",
+    "eslint": "^9.18.0",
+    "eslint-config-next": "^15.1.4",
+    "eslint-config-prettier": "^10.0.1",
     "eslint-config-react": "^1.1.7",
-    "eslint-plugin-tailwindcss": "^3.17.4",
+    "eslint-plugin-tailwindcss": "^3.17.5",
+    "lucide-react": "^0.429.0",
+    "next": "^15.1.4",
+    "next-auth": "^5.0.0-beta.25",
     "postcss": "^8",
     "prettier": "^3.4.2",
     "prettier-plugin-tailwindcss": "^0.6.9",
     "pretty-quick": "^4.0.0",
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
+    "tailwind-merge": "^2.6.0",
     "tailwindcss": "^3.4.17",
-    "typescript": "^5.7.3"
+    "tailwindcss-animate": "^1.0.7",
+    "typescript": "^5.7.3",
+    "typescript-eslint": "^8.15.0"
   }
 }
