@@ -19,9 +19,8 @@
   },
   "scripts": {
     "clean": "rm -rf .turbo node_modules chunks css react web-components",
-    "lint": "yarn lint:js & yarn lint:css",
-    "lint:js": "eslint ./src/**/*.{ts,tsx}",
-    "lint:css": "stylelint ./src/**/*.css",
+    "lint": "yarn lint:js",
+    "lint:js": "eslint ./src/**/*.{ts,tsx} --max-warnings 10",
     "typedoc": "npx typedoc --entryPointStrategy expand './src/utils/**/*.ts' --skipErrorChecking --excludeNotDocumented --plugin typedoc-plugin-markdown --plugin typedoc-github-wiki-theme --hidePageTitle true",
     "build": "tsup",
     "prepare": "npm run build",
@@ -45,21 +44,19 @@
     "@types/mocha": "^10.0.6",
     "@types/react": "^19.0.6",
     "@types/stylis": "^4.2.4",
-    "@typescript-eslint/eslint-plugin": "^6.14.0",
-    "@typescript-eslint/parser": "^6.14.0",
+    "@typescript-eslint/eslint-plugin": "^8.20.0",
+    "@typescript-eslint/parser": "^8.20.0",
     "@web/dev-server-esbuild": "^1.0.2",
     "@web/test-runner": "^0.18.1",
     "@web/test-runner-playwright": "^0.11.0",
-    "eslint": "^8.55.0",
-    "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-import": "^2.28.0",
-    "eslint-plugin-jsx-a11y": "^6.8.0",
-    "eslint-plugin-mdx": "^2.1.0",
-    "eslint-plugin-prettier": "^5.0.0",
-    "eslint-plugin-react": "^7.33.2",
-    "eslint-plugin-react-hooks": "^4.6.0",
-    "eslint-plugin-storybook": "^0.6.15",
-    "eslint-plugin-unused-imports": "^3.0.0",
+    "eslint": "^9.18.0",
+    "eslint-config-prettier": "^10.0.1",
+    "eslint-plugin-jsx-a11y": "^6.10.2",
+    "eslint-plugin-prettier": "^5.2.2",
+    "eslint-plugin-react": "^7.37.4",
+    "eslint-plugin-react-hooks": "^5.1.0",
+    "eslint-plugin-storybook": "^0.11.2",
+    "eslint-plugin-unused-imports": "^4.1.4",
     "onchange": "^7.1.0",
     "prettier": "^3.4.2",
     "react": "^19.0.0",
@@ -70,7 +67,8 @@
     "typedoc": "^0.27.6",
     "typedoc-github-wiki-theme": "^2.1.0",
     "typedoc-plugin-markdown": "^4.4.1",
-    "typescript": "^5.7.3"
+    "typescript": "^5.7.3",
+    "typescript-eslint": "^8.15.0"
   },
   "peerDependencies": {
     "@lit/react": ">=1",
