@@ -1,33 +1,86 @@
 {
-          "name": "@gyldendal/kobber-eslint",
-          "license": "MIT",
-          "version": "0.3.67",
-          "main": ".eslintrc.cjs",
-          "peerDependencies": {
-                    "@typescript-eslint/eslint-plugin": ">=6",
-                    "@typescript-eslint/parser": ">=6",
-                    "eslint": ">=8",
-                    "eslint-plugin-import": ">=2",
-                    "eslint-plugin-jsx-a11y": ">=6",
-                    "eslint-plugin-mdx": ">=2",
-                    "eslint-plugin-react": ">=7",
-                    "eslint-plugin-react-hooks": ">=4",
-                    "eslint-plugin-storybook": ">=0.6",
-                    "eslint-plugin-unused-imports": ">=3"
-          },
-          "dependencies": {
-                    "@typescript-eslint/eslint-plugin": "^6.14.0",
-                    "@typescript-eslint/parser": "^6.14.0",
-                    "eslint": "^8.55.0",
-                    "eslint-plugin-import": "^2.28.0",
-                    "eslint-plugin-jsx-a11y": "^6.8.0",
-                    "eslint-plugin-mdx": "^2.1.0",
-                    "eslint-plugin-react": "^7.33.2",
-                    "eslint-plugin-react-hooks": "^4.6.0",
-                    "eslint-plugin-storybook": "^0.6.15",
-                    "eslint-plugin-unused-imports": "^3.0.0"
-          },
-          "scripts": {
-                    "clean": "rm -rf node_modules"
-          }
+  "name": "@gyldendal/kobber-eslint",
+  "license": "MIT",
+  "version": "0.3.67",
+  "type": "module",
+  "exports": {
+    "./base": "./base.js",
+    "./next-js": "./next.js",
+    "./react": "./react.js",
+    "./ts-recommended": "./ts-recommended.js"
+  },
+  "peerDependencies": {
+    "@eslint/js": ">=9",
+    "@next/eslint-plugin-next": ">=15",
+    "@typescript-eslint/eslint-plugin": ">=8",
+    "@typescript-eslint/parser": ">=8",
+    "eslint": ">=9",
+    "eslint-config-prettier": ">=10",
+    "eslint-import-resolver-typescript": ">=3",
+    "eslint-plugin-import-x": ">=4",
+    "eslint-plugin-jsx-a11y": ">=6",
+    "eslint-plugin-react": ">=7",
+    "eslint-plugin-react-hooks": ">=5",
+    "eslint-plugin-turbo": ">=2",
+    "eslint-plugin-unused-imports": ">=4",
+    "globals": ">=15",
+    "typescript": ">=5",
+    "typescript-eslint": ">=8"
+  },
+  "peerDependenciesMeta": {
+    "@next/eslint-plugin-next": {
+      "optional": true
+    },
+    "@typescript-eslint/eslint-plugin": {
+      "optional": true
+    },
+    "@typescript-eslint/parser": {
+      "optional": true
+    },
+    "eslint-config-prettier": {
+      "optional": true
+    },
+    "eslint-import-resolver-typescript": {
+      "optional": true
+    },
+    "eslint-plugin-jsx-a11y": {
+      "optional": true
+    },
+    "eslint-plugin-react": {
+      "optional": true
+    },
+    "eslint-plugin-react-hooks": {
+      "optional": true
+    },
+    "eslint-plugin-turbo": {
+      "optional": true
+    },
+    "typescript": {
+      "optional": true
+    },
+    "typescript-eslint": {
+      "optional": true
+    }
+  },
+  "devDependencies": {
+    "@eslint/js": "^9.18.0",
+    "@next/eslint-plugin-next": "^15.1.4",
+    "@typescript-eslint/eslint-plugin": "^8.20.0",
+    "@typescript-eslint/parser": "^8.20.0",
+    "eslint": "^9.18.0",
+    "eslint-config-prettier": "^10.0.1",
+    "eslint-import-resolver-typescript": "^3.7.0",
+    "eslint-plugin-import-x": "^4.6.1",
+    "eslint-plugin-jsx-a11y": "^6.10.2",
+    "eslint-plugin-react": "^7.37.4",
+    "eslint-plugin-react-hooks": "^5.1.0",
+    "eslint-plugin-turbo": "^2.3.3",
+    "eslint-plugin-unused-imports": "^4.1.4",
+    "globals": "^15.12.0",
+    "typescript": "^5.7.3",
+    "typescript-eslint": "^8.15.0"
+  },
+  "scripts": {
+    "clean": "rm -rf node_modules"
+  }
 }
