@@ -10,9 +10,8 @@
     "web-components/**/*"
   ],
   "scripts": {
-    "lint": "yarn lint:js & yarn lint:css",
+    "lint": "yarn lint:js",
     "lint:js": "eslint ./src/**/*.{ts,tsx}",
-    "lint:css": "stylelint ./src/**/*.css",
     "prepare": "npm run build",
     "dev": "onchange \"./src/**/*.{ts,tsx,css}\" -- npm run build",
     "test": "node ./src/icons.test.js",
@@ -29,18 +28,16 @@
     "@lit/react": "^1.0.7",
     "@types/jsdom": "^21.1.7",
     "@types/react": "^19.0.2",
-    "@typescript-eslint/eslint-plugin": "^6.14.0",
-    "@typescript-eslint/parser": "^6.14.0",
-    "eslint": "^8.55.0",
-    "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-import": "^2.28.0",
-    "eslint-plugin-jsx-a11y": "^6.8.0",
-    "eslint-plugin-mdx": "^2.1.0",
-    "eslint-plugin-prettier": "^5.0.0",
-    "eslint-plugin-react": "^7.33.2",
-    "eslint-plugin-react-hooks": "^4.6.0",
-    "eslint-plugin-storybook": "^0.6.15",
-    "eslint-plugin-unused-imports": "^3.0.0",
+    "@typescript-eslint/eslint-plugin": "^8.20.0",
+    "@typescript-eslint/parser": "^8.20.0",
+    "eslint": "^9.18.0",
+    "eslint-config-prettier": "^10.0.1",
+    "eslint-plugin-jsx-a11y": "^6.10.2",
+    "eslint-plugin-prettier": "^5.2.2",
+    "eslint-plugin-react": "^7.37.4",
+    "eslint-plugin-react-hooks": "^5.1.0",
+    "eslint-plugin-storybook": "^0.11.2",
+    "eslint-plugin-unused-imports": "^4.1.4",
     "jsdom": "^24.1.0",
     "onchange": "^7.1.0",
     "prettier": "^3.4.2",
@@ -53,7 +50,8 @@
     "typedoc": "^0.27.6",
     "typedoc-github-wiki-theme": "^2.1.0",
     "typedoc-plugin-markdown": "^4.4.1",
-    "typescript": "^5.7.3"
+    "typescript": "^5.7.3",
+    "typescript-eslint": "^8.15.0"
   },
   "peerDependencies": {
     "@lit/react": ">=1",
