@@ -1,12 +1,12 @@
 /** @type {import('prettier').Config} */
-module.exports = {
-    endOfLine: "lf",
-    semi: true,
-    singleQuote: false,
-    arrowParens: "avoid",
-    tabWidth: 2,
-    printWidth: 120,
-    importOrder: ["^packages/(.*)$", "^apps/(.*)$", "^[./]"],
-    importOrderSeparation: false,
-    importOrderSortSpecifiers: true
-}
+export default {
+  endOfLine: "lf",
+  semi: true,
+  singleQuote: false,
+  arrowParens: "avoid",
+  tabWidth: 2,
+  printWidth: 120,
+  importOrder: ["^packages/(.*)$", "^apps/(.*)$", "^[./]"],
+  importOrderSeparation: false,
+  importOrderSortSpecifiers: true,
+};
