@@ -1,12 +1,8 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
-  // output: "export",
   images: {
-    // Azure does not support optimized images
     unoptimized: true,
-    dangerouslyAllowSVG: true, // ERROR (source: placeHolderImage (utils)):  The requested resource "https://placehold.co/600x400/eae0e1/691837?text=placeholder%5Cnimage" has type "image/svg+xml; charset=utf-8" but dangerouslyAllowSVG is disabled
-  },
-  images: {
+    dangerouslyAllowSVG: true,
     loader: "custom",
     loaderFile: "./lib/damImageLoader.ts",
     remotePatterns: [
